requirements.txt
ipython==8.6.0
langchain==0.0.339
llama_index==0.8.20
nebula3_python==3.4.0
networkx==3.1
numpy==1.22.4
pandas==1.5.3
pyvis==0.3.2
Requests==2.31.0
torch
tqdm==4.65.0
transformers
